chore: tooling update
Updating the dependencies of the project was causing peer issues that I didn't know how to resolve. So this is based on a new vue project with all existing stuff moved over.

There has been a breaking change in that we need to use modules instead of includes. While making the change, I noticed that most of the global_mixins.scss was entirely redundant. Not only does every browser that supports esmodules also support css-variables, but they also used different colors from the default theme. So if the fallback would be used, things would still look wrong.

Now there is no global include anymore and files that need it just import the stupid height-100 fix if needed